Navigate to order details with order object

+++ app/src/main/java/com/example/buynest/viewmodel/payment/PaymentViewModel.kt
@@ -17,7 +17,7 @@ class PaymentViewModel(
     val error: StateFlow<String?> = _error
 
     fun initiatePaymentFlow(
-        amount: Int,
+        amount: Double,
         currency: String = "usd",
         onClientSecretReady: (String) -> Unit
     ) {
